% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretHelpers.R
\name{caretModelLibCheck}
\alias{caretModelLibCheck}
\title{Check whether a caret supported model is runnable on the current machine in terms of
satisfying library dependencies}
\usage{
caretModelLibCheck(model)
}
\arguments{
\item{model}{name of the model}
}
\value{
boolean
}
\description{
Check whether a caret supported model is runnable on the current machine in terms of
satisfying library dependencies
}
