% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cooksdOutlier}
\alias{cooksdOutlier}
\title{Take a data.table and adds a column indicating whether that row is an outlier
based on cook's distance with respect to a linear regression using the provided formula}
\usage{
cooksdOutlier(dt, fml, thresh = 4, outlier.colname = "outlier")
}
\arguments{
\item{dt}{a data.table obj}

\item{fml}{formula obj}

\item{thresh}{Cooks distance threshold for outlier; if the cooks distance is greater than thresh * mean(cooks distance)
then the data point will be treated as an outlier}

\item{outlier.colname}{Column name of the new boolean column that will indicate if the row is an outlier}
}
\description{
Take a data.table and adds a column indicating whether that row is an outlier
based on cook's distance with respect to a linear regression using the provided formula
}
