% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createStyleInterval}
\alias{createStyleInterval}
\title{Helper function to create DT::styleInterval for a given data, in which
Positive numbers are mapped to shades of green, and 
Negative numbers are mapped to shades of red}
\usage{
createStyleInterval(data, num.breaks = 20)
}
\arguments{
\item{data}{a vector of possible values}

\item{num.breaks}{number of color breaks}
}
\description{
Helper function to create DT::styleInterval for a given data, in which
Positive numbers are mapped to shades of green, and 
Negative numbers are mapped to shades of red
}
