% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractUniqueColsToList}
\alias{extractUniqueColsToList}
\title{Take a data.table and return the columns with only a unique element as a separate table with
one row and the rest of the data.table separately}
\usage{
extractUniqueColsToList(dt)
}
\arguments{
\item{dt}{a data.table obj}
}
\value{
a list of two data.tables with
}
\description{
Take a data.table and return the columns with only a unique element as a separate table with
one row and the rest of the data.table separately
}
