% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_fread.R
\name{fread.s}
\alias{fread.s}
\title{shinypipe server function for the function fread}
\usage{
fread.s(input, output, session, params = list())
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{params}{list of arguments (ones that are not already included in UI)
for fread
fread.s()}
}
\description{
shinypipe server function for the function fread
}
