% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_fread.R
\name{fread.ui}
\alias{fread.ui}
\title{shinypipe UI for getting user input to
the function fread}
\usage{
fread.ui(id, label = "Input file", header = T, sep = ",")
}
\arguments{
\item{id}{namespace id (string)}

\item{label}{Label for file input}

\item{header}{Defaut value for the header}

\item{sep}{Default value for the sep
fread.ui()}
}
\description{
shinypipe UI for getting user input to
the function fread
}
