% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_formula.R
\name{s.formula}
\alias{s.formula}
\title{shinypipe server function that returns the formula}
\usage{
s.formula(input, output, session, intercept = T)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{intercept}{whether to use intercept (will be used only if intercept is NULL in ui.formula)}
}
\description{
shinypipe server function that returns the formula
}
