% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_formula.R
\name{s.formula.x}
\alias{s.formula.x}
\title{shinypipe server function for the input (independent) variable in the formula}
\usage{
s.formula.x(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
shinypipe server function for the input (independent) variable in the formula
}
