% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_formula.R
\name{s.formula.y}
\alias{s.formula.y}
\title{shinypipe server function for the response (dependent) variable in the formula}
\usage{
s.formula.y(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
shinypipe server function for the response (dependent) variable in the formula
}
