% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_fread.R
\name{s.fread}
\alias{s.fread}
\title{shinypipe server function for the function fread}
\usage{
s.fread(input, output, session, params = list())
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{params}{list of arguments for fread that are not already supplied through UI
for fread}
}
\description{
shinypipe server function for the function fread
}
