% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_params.R
\name{s.params}
\alias{s.params}
\title{shinypipe server function that returns the params
as defined in ui.params}
\usage{
s.params(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
shinypipe server function that returns the params
as defined in ui.params
}
