% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plot.R
\name{s.plot}
\alias{s.plot}
\title{shinypipe server function for creating a plot with zoom and pan functions}
\usage{
s.plot(input, output, session, plot, data, mapping = reactive(aes()),
  selected.colname = "selected", x.datatype = "numeric",
  y.datatype = "numeric")
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{plot}{'reactive' plot object that will be appended to a ggplot object}

\item{data}{'reactive' data.table to be passed to ggplot2::ggplot (cannot be NULL)}

\item{mapping}{'reactive' mapping passed to ggplot2::ggplot [Default: reactive(aes())]}

\item{selected.colname}{col name for the additional column that indicates whether the
row is selected (within the brushed region) or not [Default: "selected"]}
}
\value{
The original data with an additional column with the name given by selected.colname
which is true for points under a brush
}
\description{
shinypipe server function for creating a plot with zoom and pan functions
}
