% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_vector.R
\name{s.vector}
\alias{s.vector}
\title{shinypipe server function for getting user input for vectors}
\usage{
s.vector(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
A vector evaluated from the provided expression
}
\description{
shinypipe server function for getting user input for vectors
}
