% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{singleShadeColorInterval}
\alias{singleShadeColorInterval}
\title{Helper function to create DT::styleInterval for a given data, in which
Numbers are mapped to shades of green}
\usage{
singleShadeColorInterval(data, num.breaks = 20, reverse = F)
}
\arguments{
\item{data}{a vector of possible values}

\item{num.breaks}{number of color breaks}

\item{reverse}{reverse the color scale [default: F]}
}
\description{
Helper function to create DT::styleInterval for a given data, in which
Numbers are mapped to shades of green
}
