% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_caretModelList.R
\name{ui.caretModelList}
\alias{ui.caretModelList}
\title{shinypipe UI for creating a data frame that can potentially be used
as a tuning grid for caret::train}
\usage{
ui.caretModelList(id, models, selected = NULL)
}
\arguments{
\item{id}{namespace id (string)}

\item{models}{reactive list of models}
}
\description{
shinypipe UI for creating a data frame that can potentially be used
as a tuning grid for caret::train
}
