% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_flexOutput.R
\name{ui.flexOutput}
\alias{ui.flexOutput}
\title{shinypipe UI for creating a flexible output UI}
\usage{
ui.flexOutput(id, editable = T, plotHeight = 400)
}
\arguments{
\item{id}{namespace id (string)}

\item{editable}{whether the user can change the output type and output expression [Default: F]}

\item{plotHeight}{plot height when plot type is chosen [Default: 400]}
}
\description{
shinypipe UI for creating a flexible output UI
}
