% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_fread.R
\name{ui.fread}
\alias{ui.fread}
\title{shinypipe UI for getting user input to
the function fread}
\usage{
ui.fread(id, fileIn = list(label = "Input file"), sep = ",",
  header = T)
}
\arguments{
\item{id}{namespace id (string)}

\item{fileIn}{list of params for shiny::fileInput to select a file}

\item{sep}{Selected value for the separator (default: ",")
If user doesn't need this control, set it to NULL, and control the parameter through params in s.fread}

\item{header}{Whether file has headers (default: T)
If user doesn't need this control, set it to NULL, and control the parameter through params in s.fread}
}
\description{
shinypipe UI for getting user input to
the function fread
}
