% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_listOfVectors.R
\name{ui.listOfVectors}
\alias{ui.listOfVectors}
\title{shinypipe UI for creating a data frame that can potentially be used
as a tuning grid for caret::train}
\usage{
ui.listOfVectors(id, table.params, showValues = F)
}
\arguments{
\item{id}{namespace id (string)}

\item{table.params}{reactive data frame of parameters with the following columns:
parameter - name of the parameter
class - One of "numeric", "character" or "logical" [Default: "numeric"]
label - [Default: same as parameter]
value - Expression value to be populated [Default: ""]}

\item{showValues}{show the evaluated vector next to each row}
}
\description{
shinypipe UI for creating a data frame that can potentially be used
as a tuning grid for caret::train
}
