% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plot.R
\name{ui.plot}
\alias{ui.plot}
\title{shinypipe UI for creating a plot with zoom and pan functions}
\usage{
ui.plot(id, height = 400)
}
\arguments{
\item{id}{namespace id (string)}

\item{height}{height of the widget}
}
\description{
shinypipe UI for creating a plot with zoom and pan functions
}
