% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_caretModel.R
\name{s.caretModel}
\alias{s.caretModel}
\title{shinypipe server function for returning a data frame that can potentially be used
as a tuning grid for caret::train}
\usage{
s.caretModel(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{return}{reactive list}
}
\description{
shinypipe server function for returning a data frame that can potentially be used
as a tuning grid for caret::train
}
