% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_flexOutput.R
\name{s.flexOutput}
\alias{s.flexOutput}
\title{shinypipe server function for returning a data frame that can potentially be used
as a tuning grid for caret::train}
\usage{
s.flexOutput(input, output, session, expr = "", type = "Print")
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{expr}{The expression to be evaluated as a string, or its starting value if editable}

\item{type}{The type of UI, or its starting value if editable (one of "Print", "Table", or "Plot")}
}
\value{
a reactive list of type and expr used
}
\description{
shinypipe server function for returning a data frame that can potentially be used
as a tuning grid for caret::train
}
