% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_caretModel.R
\name{ui.caretModel}
\alias{ui.caretModel}
\title{shinypipe UI for creating a data frame that can potentially be used
as a tuning grid for caret::train}
\usage{
ui.caretModel(id, table.params)
}
\arguments{
\item{id}{namespace id (string)}

\item{table.params}{reactive data frame of parameters with three columns:
parameter, class and label similar to that provided by the parameter item in the
object returned by caret::getModelInfo}
}
\description{
shinypipe UI for creating a data frame that can potentially be used
as a tuning grid for caret::train
}
