% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_vector.R
\name{ui.vector}
\alias{ui.vector}
\title{shinypipe UI for for getting user input for vectors}
\usage{
ui.vector(id, type = "numeric", label = "Expression", value = "",
  showValues = T)
}
\arguments{
\item{id}{namespace id (string)}

\item{type}{One of "numeric", "character" or "logical". For now,
no validation is provided between character and numeric types and the user is just expected
to enter the correct types}

\item{label}{label for the expression to be evaluated as a vector}

\item{value}{initial value for the expression to be evaluated as a vector}
}
\description{
shinypipe UI for for getting user input for vectors
}
